+++ framework/core/js/src/forum/forum.ts
@@ -22,6 +22,7 @@ import './components/HeaderPrimary';
 import './components/PostEdited';
 import './components/IndexPage';
 import './components/DiscussionRenamedNotification';
+import './components/DiscussionsSearchSource';
 import './components/HeaderSecondary';
 import './components/DiscussionList';
 import './components/AvatarEditor';
@@ -31,6 +32,7 @@ import './components/NotificationsDropdown';
 import './components/UserPage';
 import './components/PostUser';
 import './components/UserCard';
+import './components/UsersSearchSource';
 import './components/PostPreview';
 import './components/EventPost';
 import './components/DiscussionHero';
