+++ framework/core/less/forum/DiscussionListItem.less
@@ -37,12 +37,6 @@
   position: absolute;
   top: 0;
   left: -2px;
-
-  .Badge {
-    margin-left: -10px;
-    position: relative;
-    pointer-events: auto;
-  }
 }
 .DiscussionListItem-title {
   margin: 0 0 3px;
@@ -132,12 +126,6 @@
     &:empty {
       display: none;
     }
-  }
-
-  .Badge {
-    --size: 18px;
-    box-shadow: none;
-    margin-left: -8px;
   }
 
   .DiscussionListItem-title {
