+++ framework/core/less/forum/Post.less
@@ -84,12 +84,6 @@
   text-align: right;
   white-space: nowrap;
   pointer-events: none;
-
-  .Badge {
-    margin-left: -10px;
-    position: relative;
-    pointer-events: auto;
-  }
 }
 
 .Post-body {
